+++ mmodel/model.py
@@ -23,10 +23,10 @@ class Model(EditMixin, ReprMixin):
         the keyword arguments.
     :param dict handler_kwargs: keyword arguments for the handler class.
     :param list modifiers: modifiers used for the whole graph model executable.
-    :param list returns: The order of model returns; defaults to the
-        topological search.
+    :param list returns: If not suplied, the returns are the returns of the terminal
+        roots. The order of model returns defaults to the topological order.
     :param str doc: model docstring
-    :param dict defaults: default values for the model signature.
+    :param dict param_defaults: default values for the model signature.
     :param bool kw_only: whether to convert signature to keyword-only signature
     """
 
@@ -38,7 +38,7 @@ class Model(EditMixin, ReprMixin):
         handler_kwargs: dict = None,
         modifiers: list = None,
         returns: list = None,
-        defaults: dict = None,
+        param_defaults: dict = None,
         doc: str = "",
         **kwargs,
     ):
@@ -50,7 +50,7 @@ class Model(EditMixin, ReprMixin):
         self._modifiers = modifiers or list()
         self.handler = handler
         self._handler_kwargs = handler_kwargs or {}
-        self._defaults = defaults or {}
+        self._param_defaults = param_defaults or {}
         self.doc = self.__doc__ = doc
 
         # update the kwargs
@@ -65,7 +65,7 @@ class Model(EditMixin, ReprMixin):
         self.model_func = modify_func(self._runner, self._modifiers)
         # apply defaults to model_func
         self.model_func.__signature__ = restructure_signature(
-            signature(self.model_func), self._defaults
+            signature(self.model_func), self._param_defaults
         )
 
     @property
@@ -100,9 +100,9 @@ class Model(EditMixin, ReprMixin):
         return self._modifiers.copy()
 
     @property
-    def defaults(self):
+    def param_defaults(self):
         """Shallow copy of the defaults."""
-        return self._defaults.copy()
+        return self._param_defaults.copy()
 
     @property
     def handler_kwargs(self):